+++ alpha/apps/kaltura/lib/kUploadTokenMgr.php
@@ -7,6 +7,8 @@ class kUploadTokenMgr
 	const MAX_APPEND_TIME = 5;
 	const MAX_CHUNKS_WAITING_FOR_CONCAT_ALLOWED = 1000;
 	const CHUNK_SIZE = 102400;
+	const ENTRY = 'entry';
+	const FLAVOR = 'flavorAsset';
 
 	/**
 	 * @var UploadToken
@@ -124,8 +126,33 @@ class kUploadTokenMgr
 		$allowedStatuses = array(UploadToken::UPLOAD_TOKEN_PENDING, UploadToken::UPLOAD_TOKEN_PARTIAL_UPLOAD);
 		if (!in_array($this->_uploadToken->getStatus(), $allowedStatuses, true))
 			throw new kUploadTokenException("Invalid upload token status", kUploadTokenException::UPLOAD_TOKEN_INVALID_STATUS);
-
+		
 		$this->updateFileName($fileData);
+		
+		$entryUpdated = false;
+		$entry = $this->getAttachedEntry();
+		if ($entry && $entry->getType() == entryType::MEDIA_CLIP && !$entry->getMediaType())
+		{
+			$mediaType = $this->getMediaType();
+			if ($mediaType)
+			{
+				$entry->setMediaType($this->getMediaType());
+				$entryUpdated = true;
+			}
+			
+			if($entryUpdated)
+			{
+				$entry->save();
+			}
+			
+			if($mediaType && $mediaType == KalturaMediaType::IMAGE)
+			{
+				$contentResource = new KalturaUploadedFileTokenResource();
+				$contentResource->token = $this->_uploadToken->getId();
+				$dbContentResource = $contentResource->toObject();
+				$dbContentResource->attachCreatedObject($entry);
+			}
+		}
 		
 		try
 		{
@@ -168,7 +195,7 @@ class kUploadTokenMgr
 		}
 		else
 		{
-			//We return null file size when we want to faile the upload since it reached max chunks waiting for concat
+			//We return null file size when we want to fail the upload since it reached max chunks waiting for concat
 			$this->_uploadToken->setStatus(!is_null($fileSize) ? UploadToken::UPLOAD_TOKEN_PARTIAL_UPLOAD : UploadToken::UPLOAD_TOKEN_ERROR);
 		}
 		
@@ -176,6 +203,48 @@ class kUploadTokenMgr
 		$this->_uploadToken->setDc(kDataCenterMgr::getCurrentDcId());
 		
 		$this->_uploadToken->save();
+	}
+	
+	protected function getAttachedEntry()
+	{
+		$entry = null;
+		if ($this->_uploadToken->getObjectId() && $this->_uploadToken->getObjectType())
+		{
+			switch ($this->_uploadToken->getObjectType())
+			{
+				case self::ENTRY:
+				{
+					$entry = entryPeer::retrieveByPK($this->_uploadToken->getObjectId());
+					break;
+				}
+				case self::FLAVOR:
+				{
+					$flavorAsset = assetPeer::retrieveById($this->_uploadToken->getObjectId());
+					if ($flavorAsset && $flavorAsset->getType() == assetType::FLAVOR && $flavorAsset->getEntryId())
+					{
+						$entry = entryPeer::retrieveByPK($flavorAsset->getEntryId());
+					}
+					break;
+				}
+			}
+		}
+		return $entry;
+	}
+	
+	protected function getMediaType()
+	{
+		if(is_null($this->_uploadToken))
+		{
+			return null;
+		}
+		
+		$fileName = $this->_uploadToken->getFileName();
+		if(!$fileName)
+		{
+			return null;
+		}
+		
+		return myFileUploadService::getMediaTypeFromFileExt(pathinfo($fileName, PATHINFO_EXTENSION));
 	}
 	
 	/**