+++ alpha/lib/enums/ExportObjectType.php
@@ -8,4 +8,5 @@ interface ExportObjectType extends BaseEnum
 {
 	const USER = 1;
 	const ENTRY = 2;
+	const CATEGORY = 3;
 }