+++ plugins/scheduled_task/lib/MediaRepurposingHandler.php
@@ -6,6 +6,7 @@
  */
 class MediaRepurposingHandler implements kObjectDataChangedEventConsumer
 {
+    	const MRP_IDS_TO_EXCLUDE_RESET_ON_METADATA_UPDATE = "MRP_ids_to_exclude_reset_on_metadata_update";
 
 	public static function enableMrPermission($partnerId)
 	{
@@ -70,19 +71,25 @@ class MediaRepurposingHandler implements kObjectDataChangedEventConsumer
 
 	private function getMRPWithMetadataSearchByProfile($partnerId, $metadataProfileId)
 	{
-		$allMediaRepurposingProfilesOnPartner = ScheduledTaskProfilePeer::retrieveBySystemName(MediaRepurposingUtils::MEDIA_REPURPOSING_SYSTEM_NAME, $partnerId);
-		$mediaRepurposingProfilesWithSearchOnGivenMetadataId = array();
-		foreach($allMediaRepurposingProfilesOnPartner as $mediaRepurposingProfile) {
-			/* @var $mediaRepurposingProfile ScheduledTaskProfile*/
-			if (!$mediaRepurposingProfile->getObjectFilter() || !$mediaRepurposingProfile->getObjectFilter()->getAdvancedSearch())
-				continue;
-			$items = $mediaRepurposingProfile->getObjectFilter()->getAdvancedSearch()->getItems(); // always have items in advance search because that how the MR is build
-			foreach($items as $item) {
-				if ($item instanceof MetadataSearchFilter && $item->getMetadataProfileId() == $metadataProfileId)
-					$mediaRepurposingProfilesWithSearchOnGivenMetadataId[] = $mediaRepurposingProfile->getId();
-			}
-		}
-		return $mediaRepurposingProfilesWithSearchOnGivenMetadataId;
+        	$allMediaRepurposingProfilesOnPartner = ScheduledTaskProfilePeer::retrieveBySystemName(MediaRepurposingUtils::MEDIA_REPURPOSING_SYSTEM_NAME, $partnerId);
+        	$mediaRepurposingProfilesWithSearchOnGivenMetadataId = array();
+        	$MRProfilesToExclude = kConf::get(self::MRP_IDS_TO_EXCLUDE_RESET_ON_METADATA_UPDATE, kConfMapNames::RUNTIME_CONFIG, array());
+
+        	foreach($allMediaRepurposingProfilesOnPartner as $mediaRepurposingProfile)
+        	{
+        	    /* @var $mediaRepurposingProfile ScheduledTaskProfile*/
+        	    if(in_array($mediaRepurposingProfile->getId(), $MRProfilesToExclude) || !$mediaRepurposingProfile->getObjectFilter() || !$mediaRepurposingProfile->getObjectFilter()->getAdvancedSearch())
+        	        continue;
+        	    $items = $mediaRepurposingProfile->getObjectFilter()->getAdvancedSearch()->getItems(); // always have items in advance search because that how the MR is build
+        	    foreach($items as $item)
+        	    {
+        	        if($item instanceof MetadataSearchFilter && $item->getMetadataProfileId() == $metadataProfileId)
+        	        {
+        	            $mediaRepurposingProfilesWithSearchOnGivenMetadataId[] = $mediaRepurposingProfile->getId();
+        	        }
+        	    }
+        	}
+        	return $mediaRepurposingProfilesWithSearchOnGivenMetadataId;
 	}
 
 	private function getMediaRepuposingMetadataProfileId($partnerId)